requirements.txt
torch
torchvision
torch_ac
minigrid
gymnasium
numpy
matplotlib
pyyaml
pillow
